Quite hacky way to make things work using deferred callbacks.
When using deferred functions, in order to make synchronous interfaces
work, we need to call _step() instead of asyncore.loop() during the
wait cycles of make_sync. But _step() is a function of Service, so
we need the make_sync decorator to be aware of the running Service,
that is not available in the global context.

Unfortunately, decorators in classes are applied at class build time,
not at instatiation time, so I found no way to apply a decorator
with an instance as parameter other than manually do the decoration
in __init__(). This leads to a fair number of levels of indirection
that I don't like very much. Moreover, this impementation still has
shortcomings, so I think we should discuss it.

So far, it'll go in a separate branch for evaluating.